Physics Stuff
There's two player objects now and they share a position, but not
forces. You'll notice that if only one of the instances is standing on a
block then you can't jump nearly as high because one character is doing
the jumping for both. The flat red block is a spring, if you phase
almost entirely into the top world the bottom character will be light
enough for the spring to throw you over the last wall.
The current phase level is printed to the console, between -1 (entirely
bottom world) and 1 (entirely top world). Hopefully that all makes
sense.