Pickup Buttons
There is now a pick up button for each world.
Top world - Pickup1 - f
Bottom World - Pickup 2 - g

New Public Variable for Pickup script - input string: use Pickup1 or
Pickup2
Input string determines what input the script on the box is looking for
and finds the nearest the nearest box with the right tag for the player
to pick up.

Thought - attach the pickup script to the player and just have all the
boxes correctly tagged.

Updated Tags to have Pickup1 and Pickup2
Updated Input to have Pickup1 and Pickup2
Updated boxes with the needed tags and variables.

Note: Initial testing shows the script works correctly in Level1, but
the player is unable to pick up anything in Level2.