Another Quick Fix on Pickup
Added more checks in PickupKey1 and PickupKey2

Found an issue where if a box was wrongly tagged/scripted, i.e. tagged
Pickup2 but had the script PickupMouse1. NullReferenceExceptions would
be thrown as PickupKey2 would be looking for the script PickupMouse2 on
at box, which of course doesn't exist.

Added more checks the prevent the exception being thrown, however if
things are wrongly tagged/scripted pickup wont work as intended anyway.