Quick Fix On Pickup
Found an issue. There was a box in the top world tagged as "Pickup2" so
it would check for the script PickupMouse2. But that box had
PickupMouse1 script instead. Thus NullReferenceExceptions were thrown.

Fixed the tagging and added more checks to prevent the exceptions.
However if things are not tagged/scripted correctly Pickup wont work as
intended.