Sounds!
Renamed the Audio Files.

Added some initial places in levels for sound to play when the player is
close enough to them. i.e. puzzle hints

Added sound when Phasing is not safe.

Currently having issue with adding sound when the player is stunned,
multiple different attempts, different errors and problems each time.