Level 3 Draft
Added the initial draft of level 3, with the 4 puzzles we discussed.

Some notes with the platforms:
Added a part in the code that stops the platform (removes all forces
currently acting upon it) when it gets really close the the point where
it is suppose to change direction. This is done to stop the platform
over shooting the point and some times colliding and getting stuck on
boxes when it over shoots + elasticity.

When the platforms bob up and down too much (at high speeds) the
platform can get stuck on the boxes.

Careful: platforms can throw the player off it when it changes
directions due to the sudden change in forces.

Part 1 - 1 moving platform:
No real issues in testing

Part 2 - 2 platforms in same world:
current implementation of moving platforms makes this one tricky, if you
get on one of the platforms, the platforms get out of sync, making
jumping to the other difficult.

Part 3 - 1 Platform, use change in weight to avoid the hanging pillar:
Current implementation of weight/platform doesn't add much effect to
changing the y position, and the elasticity just makes it bob up and
down a bit more.

Part 4 - 2 Platforms, 1 in each world:
No real issue, difficulty is there. Can be done without jump to the next
platform just phasing.

Need discussion on where to take the puzzles next.

Current Error: Can get the exit door to work to exit the level.