Level 3 Improvement
Ryan: I've fixed part 1 and the first platform in part 4 to demonstrate
the sort of settings you need. VForce counteracts gravity, so you will
usually want it set to mass*15 unless you're going for a bouncy effect
with a platform. Mass should be set higher than 1 because if it's 1 then
the platform weighs less than the player. Elasticity controls how
strongly the platform will fight against the players weight. You will
also probably always want drag values that aren't 0.

For the first platform in part 4 I set the mass to 5, the VForce to 75
to cancel out gravity, the Elasticity to a low value so that the
platform responds a lot to changing player weight and drag to 0.5 so
that it responds to those forces in a reasonable way.

You just need to find the right values to get the effect you want.

I would also recommend at least modifying the script so that the
platforms decelerate as they approach their destination, rather than
stopping suddenly. Preferably you would add a timing feature to the
script so that a platform will accelerate or decelerate to reach its
destination at the right time, so that part 2 works properly.