Sound Update
Narration now plays at the Camera1. IE doesn't change volume if the
player moves (the previous version).

Camera 1 is now a prefab.

Edited Player script to work with Camera1.

New script: PlayAtCamera. When the player gets near the game object, the
selected audio clip will play at the camera. Provided the clip is on the
camera as an audio source.