Sound Rework
At Stephen's suggestion I've redone how sound works for narration.

When the player is in proximity to the sound object with the
ProximityPlay script attached to it. The script calls the
SoundPlay(AudioClip) function in CameraPlay and uses PlayOneShot to play
the AudioClip given in the ProximityPlay script.

Warnings like "Phasing Into Solid Matter" are still dealt with by the
previous method with the camera. This is due to the isPlaying variable
the AudioComponent has, which when used with an IF statement, stops the
spamming of the warning messages.

Brought Updated Level 4 with sound, prefabs and camera anchors.