Pause Feature
Implemented a Pause feature.

The pause script is just attached to an empty Object in each level
called "Pause"

The script sets Time.timeScale to 0 when paused. and sets it back to 1
to unpause.

Current escape Key is ESC. since it is a special button it is hardcoded
in the Script and not dealt with by the input manager.

Edited Player to not phase or adjust the time scale via phasing while
paused. Just added checks for a boolean called "isPaused". Pause script
has a function which returns isPaused called GetIsPaused().