Guard Animations!
Guard Animations are in.

Walking, jumping and Tazering.

Had some issues with the explode and squish animations, so i left those
out for now.

Changed prefabs to reflect the new guards.

3 different guards:
Shooting Guards - the patrolling kind in level 2
Crushed Guards - the ones that get crushed in level 1
Static Guards - the static guards in level 3 and level 5

Reorganised the animations, deleted some of the old ones.

Made the needed changes across the levels to use the new prefab guards.