Intercom Lines + End
I've revised the intercom trigger prefab. If it has a clip associated
with it that will be added to the intercom queue, if no clip then it
adds an idle clip to the queue.
Gizmos have been colour coded. Green is an idle intercom. Blue is an
intercom line. White is camera anchor.
I've added a trigger at the end of Level 5 that plays the last intercom
lines and then cuts to the "Credits" scene. The Credits scene only has
music at the moment.