Elevator Animations!
Elevator Animations are in.

New prefabs for the elevators at the start of a level and at the end of
a level.

positions of the new elevators are a little bit different from the old
elevators, the new positions may not be perfect.

tweaked the starting position of the Player to with the addition of the
new elevators.